Initial scaffolding using wp-cli to setup plugin structure with Unit Test support.
At this point, we have a rootQuery `fieldGroups` that will return existing act field groups and their properties.

Fields on Field Groups are returned as a list of fieldUnion, as the field types aren’t known until query time.

An example query that could be run at this time is:

```
{
  fieldGroups{
    id
    title
    key
    fieldGroupId
    locationRules{
      param
      operator
      value
    }
    fields{
      ...on textField{
        id
        name
        instructions
      }
    }
  }
}
```